Clarify comments about LitePCIeDMASynchronizer
The new mode allows synchronization of TX streams to RX streams upon a PPS event, or only PPS with TX and RX streams unsynchronized. This clarifies that in "PPS-only" mode, both TX and RX streams can be "flowgated" by the synchronizer module, but they may remain unsynced to each other.