litepcie_gen: Limit IRQ rate to 1MHz on Ultrascale(+) to prevent IRQ stall issue.
On Ultrascale(+) IRQ stop being transmitted without this. The issue still need to be
closely investigated but this IRQ rate limitation seems to prevent it.

Since interface with the PHY will be soon reworked, it's not useful to investigate now,
we'll investigate after the rework if the issue is still present.